Document re-export of archive helpers
